substituindo tipos e ajustando retorno do aStar

+++ agents/cpp/support.cpp
@@ -1,6 +1,6 @@
 #include "support.hpp"
 
-float Node::calculateHeuristic(Node target) {
+int Node::calculateHeuristic(Node target) {
   Support support;
   return support.manhattanDistance(this->coordinates, target.coordinates);
 }
@@ -13,7 +13,7 @@ bool Coordinates::isEqual(Coordinates point) {
   return (this->x == point.x) && (this->y == point.y);
 }
 
-float Support::manhattanDistance(Coordinates start, Coordinates target) {
+int Support::manhattanDistance(Coordinates start, Coordinates target) {
   return abs(start.x - target.x) + abs(start.y - target.y);
 }
 
@@ -23,7 +23,7 @@ bool Support::containsInPriorityQueue(Node node, priority_queue<Node, vector<Nod
     priorityQueue.pop();
     if (n.coordinates.isEqual(node.coordinates))
       return true;
-    }
+  }
   return false;
 }
